smtp_relay/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
aiosmtpd==1.4.4
python-multipart==0.0.6
dnspython==2.4.2
